Wasmtime: Move `Global` and `Table` to `externals` submodules
Just mechanical code motion, not functional changes.
